magazine/requirements.txt
Django==1.4.10
South==0.8.4
django-articles==2.4.1
ipython==1.2.0
wsgiref==0.1.2
gunicorn==18.0
django-tinymce==1.5.2
